// apps/dashboard/src/app/features/products/models/product-query.model.ts
export type ProductSortBy =
  | 'title' | 'price' | 'rating' | 'createdAt'| 'bestSelling' | 'mostPopular';

export type SortOrder = 'asc' | 'desc';


export interface ProductQueryParams {
  page?: number;
  limit?: number;
  categoryId?: string;
  subCategoryId?: string;
  occasionId?: string;
  minPrice?: number;
  maxPrice?: number;
  minRating?: number;
  sortBy?: ProductSortBy;
  sortOrder?: SortOrder;
  search?: string;
}